Use glib's GHashTable instead of persistent BDB
I'd still like to rework this into something a bit nicer, but this patch takes
the heat off and allows me to finish moving Graphite to eqiad without feeling
like I'm moving an old problem to a new data center. It chugs along at 30-40%
CPU compared to the previous version which maxed out and dropped data.

Dumping stats in a separate thread would be nice but apparently not critical.
I removed the standalone 'exporter' tool as there is no longer a database file
for it to read.

Adds a dependency on libglib2.0-dev. The version in apt is 2.32, so the
relevant docs are: <https://developer.gnome.org/glib/2.32/glib-Hash-Tables.html>.

Change-Id: I47e4ed0bb2ec49cb4472716b21abb01526d0ee49